feat: TeamMember 엔티티 설계
