+++ src/main/java/store/dropthebeatbox/app/domain/mapping/Friends.java
@@ -1,0 +1,26 @@
+package store.dropthebeatbox.app.domain.mapping;
+
+import lombok.*;
+import store.dropthebeatbox.app.domain.Member;
+
+import javax.persistence.*;
+
+@Entity
+@Builder
+@Getter
+@AllArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Friends {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "owner_id")
+    private Member owner;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "target_id")
+    private Member target;
+}
