+++ src/main/java/store/dropthebeatbox/app/domain/mapping/TeamMember.java
@@ -1,6 +1,8 @@
-package store.dropthebeatbox.app.domain;
+package store.dropthebeatbox.app.domain.mapping;
 
 import lombok.*;
+import store.dropthebeatbox.app.domain.Member;
+import store.dropthebeatbox.app.domain.Team;
 
 import javax.persistence.*;
 
@@ -9,17 +11,17 @@ import javax.persistence.*;
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Friends {
+public class TeamMember {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "owner_id")
-    private Member owner;
+    @JoinColumn(name = "member_id")
+    private Member member;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "target_id")
-    private Member target;
+    @JoinColumn(name = "team_id")
+    private Team team;
 }
